+++ checkpoint/orbax/checkpoint/type_handlers.py
@@ -396,10 +396,12 @@ class SaveArgs:
 
   def __post_init__(self):
     if self.aggregate:
-      logging.warning(
+      logging.log_every_n_seconds(
+          logging.WARNING,
           'SaveArgs.aggregate is deprecated, please use custom TypeHandler'
           ' (https://orbax.readthedocs.io/en/latest/custom_handlers.html#typehandler)'
-          ' or contact Orbax team to migrate before May 1st, 2024.'
+          ' or contact Orbax team to migrate before May 1st, 2024.',
+          n_seconds=12 * 60 * 60,  # once every 12 hours
       )