+++ checkpoint/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add `reload` method in `CheckpointManager` which resets internal properties.
+- Add checkpoint announcements RTD page.
 
 ### Changed
 - Deprecate `read` option in `all_steps`.