+++ checkpoint/CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Deprecate `read` option in `all_steps`.
 - Deprecate `SaveArgs.aggregate`.
+- Update copyright to use 2024. Also, allow notebook cells to continue if last
+ cell raises error.
 
 ## [0.5.2] - 2024-01-26